requirements.txt
google-genai
pydantic
pydantic-ai
python-dotenv
streamlit